+++ keepalived/check/check_daemon.c
@@ -695,7 +695,7 @@ start_check_child(void)
 	prctl(PR_SET_PDEATHSIG, SIGTERM);
 
 	/* Check our parent hasn't already changed since the fork */
-	if (main_pid != getppid())
+	if (main_pid != 0 && main_pid != getppid())
 		kill(getpid(), SIGTERM);
 
 	prog_type = PROG_TYPE_CHECKER;
